+++ src/test/java/org/openlmis/stockmanagement/testutils/StockCardLineItemDataBuilder.java
@@ -112,8 +112,18 @@ public class StockCardLineItemDataBuilder {
     return this;
   }
 
+  public StockCardLineItemDataBuilder withId(UUID id) {
+    this.id = id;
+    return this;
+  }
+
   public StockCardLineItemDataBuilder withOriginEvent(StockEvent event) {
     originEvent = event;
+    return this;
+  }
+
+  public StockCardLineItemDataBuilder withDocumentNumber(String documentNumber) {
+    this.documentNumber = documentNumber;
     return this;
   }